+++ models/ad_reporting__monthly_campaign_region_report.sql
@@ -1,50 +1,23 @@
 {% set include_list = [] %}
-{% set exclude_list = ['amazon_ads', 'apple_search_ads', 'google_ads', 'reddit_ads', 'tiktok_ads'] %}
+{% do include_list.append('facebook_ads') if var('facebook_ads__using_demographics_region', false) %}
+{% do include_list.append('linkedin_ads') if var('linkedin_ads__using_geo', true) and var('linkedin_ads__using_monthly_ad_analytics_by_member_region', true) %}
+{% do include_list.append('microsoft_ads') if var('microsoft_ads__using_geographic_daily_report', false) %}
+{% do include_list.append('pinterest_ads') if var('pinterest__using_pin_promotion_targeting_report', true) and var('pinterest__using_targeting_geo_region', true) %}
+{% do include_list.append('snapchat_ads') if var('snapchat_ads__using_campaign_region_report', false) %}
+{% do include_list.append('twitter_ads') if var('twitter_ads__using_campaign_regions_report', false) %}
 
-{% if var('facebook_ads__using_demographics_region', false) %}
-    {% do include_list.append('facebook_ads') %}
+-- Used for logging macro behavior, can be removed after PR review
+{{ log("include_list before filtering for region report: " ~ include_list, info=True) }}
+
+{% if include_list | length > 0 %}
+    {% set enabled_packages = get_enabled_packages(include=include_list) %}
 {% else %}
-    {% do exclude_list.append('facebook_ads') %}
+    {% set enabled_packages = [] %}
 {% endif %}
 
-{% if var('linkedin_ads__using_geo', true) and var('linkedin_ads__using_monthly_ad_analytics_by_member_region', true) %}
-    {% do include_list.append('linkedin_ads') %}
-{% else %}
-    {% do exclude_list.append('linkedin_ads') %}
-{% endif %}
-
-{% if var('microsoft_ads__using_geographic_daily_report', false) %}
-    {% do include_list.append('microsoft_ads') %}
-{% else %}
-    {% do exclude_list.append('microsoft_ads') %}
-{% endif %}
-
-{% if var('pinterest__using_pin_promotion_targeting_report', true) and var('pinterest__using_targeting_geo_region', true) %}
-    {% do include_list.append('pinterest_ads') %}
-{% else %}
-    {% do exclude_list.append('pinterest_ads') %}
-{% endif %}
-
-{% if var('snapchat_ads__using_campaign_region_report', false) %}
-    {% do include_list.append('snapchat_ads') %}
-{% else %}
-    {% do exclude_list.append('snapchat_ads') %}
-{% endif %}
-
-{% if var('twitter_ads__using_campaign_regions_report', false) %}
-    {% do include_list.append('twitter_ads') %}
-{% else %}
-    {% do exclude_list.append('twitter_ads') %}
-{% endif %}
-
-{# Log which packages were added to which list #}
-{{ log("\n Region report - include_list before filtering: " ~ include_list, info=True) }}
-{{ log("\n Region report - exclude_list before filtering: " ~ exclude_list, info=True) }}
-
-{% set enabled_packages = get_enabled_packages(exclude=exclude_list)%}
-
-{# Log which packages survived global enablement filtering #}
-{{ log("\n Region report - enabled_packages after macro filtering: " ~ enabled_packages, info=True) }}
+-- Used for logging macro behavior, can be removed after PR review
+{{ log("enabled_packages after macro filtering for region report: " ~ enabled_packages, info=True) }}
+{{ log("Model enabled for region report? " ~ is_enabled(enabled_packages), info=True) }}
 
 {{ config(enabled=is_enabled(enabled_packages)) }}