+++ models/intermediate/int_ad_reporting__monthly_campaign_country_report.sql
@@ -1,62 +1,26 @@
 {% set include_list = [] %}
-{% set exclude_list = ['amazon_ads', 'apple_search_ads', 'google_ads'] %}
+{% do include_list.append('facebook_ads') if var('facebook_ads__using_demographics_country', false) %}
+{% do include_list.append('linkedin_ads') if var('linkedin_ads__using_geo', true) and var('linkedin_ads__using_monthly_ad_analytics_by_member_country', true) %}
+{% do include_list.append('microsoft_ads') if var('microsoft_ads__using_geographic_daily_report', false) %}
+{% do include_list.append('pinterest_ads') if var('pinterest__using_pin_promotion_targeting_report', true) and var('pinterest__using_targeting_geo', true) %}
+{% do include_list.append('reddit_ads') if var('reddit_ads__using_campaign_country_report', true) %}
+{% do include_list.append('snapchat_ads') if var('snapchat_ads__using_campaign_country_report', false) %}
+{% do include_list.append('tiktok_ads') if var('tiktok_ads__using_campaign_country_report', true) %}
+{% do include_list.append('twitter_ads') if var('twitter_ads__using_campaign_locations_report', false) %}
 
-{% if var('facebook_ads__using_demographics_country', false) %}
-    {% do include_list.append('facebook_ads') %}
+-- Used for logging macro behavior, can be removed after PR review
+{{ log("include_list before filtering for intermediate country report: " ~ include_list, info=True) }}
+
+{% if include_list | length > 0 %}
+    {% set enabled_packages = get_enabled_packages(include=include_list) %}
 {% else %}
-    {% do exclude_list.append('facebook_ads') %}
+    {% set enabled_packages = [] %}
 {% endif %}
 
-{% if var('linkedin_ads__using_geo', true) and var('linkedin_ads__using_monthly_ad_analytics_by_member_country', true) %}
-    {% do include_list.append('linkedin_ads') %}
-{% else %}
-    {% do exclude_list.append('linkedin_ads') %}
-{% endif %}
+-- Used for logging macro behavior, can be removed after PR review
+{{ log("enabled_packages after macro filtering for intermediate country report: " ~ enabled_packages, info=True) }}
+{{ log("Model enabled for intermediate country report? " ~ is_enabled(enabled_packages), info=True) }}
 
-{% if var('microsoft_ads__using_geographic_daily_report', false) %}
-    {% do include_list.append('microsoft_ads') %}
-{% else %}
-    {% do exclude_list.append('microsoft_ads') %}
-{% endif %}
-
-{% if var('pinterest__using_pin_promotion_targeting_report', true) and var('pinterest__using_targeting_geo', true) %}
-    {% do include_list.append('pinterest_ads') %}
-{% else %}
-    {% do exclude_list.append('pinterest_ads') %}
-{% endif %}
-
-{% if var('reddit_ads__using_campaign_country_report', true) %}
-    {% do include_list.append('reddit_ads') %}
-{% else %}
-    {% do exclude_list.append('reddit_ads') %}
-{% endif %}
-
-{% if var('snapchat_ads__using_campaign_country_report', false) %}
-    {% do include_list.append('snapchat_ads') %}
-{% else %}
-    {% do exclude_list.append('snapchat_ads') %}
-{% endif %}
-
-{% if var('tiktok_ads__using_campaign_country_report', true) %}
-    {% do include_list.append('tiktok_ads') %}
-{% else %}
-    {% do exclude_list.append('tiktok_ads') %}
-{% endif %}
-
-{% if var('twitter_ads__using_campaign_locations_report', false) %}
-    {% do include_list.append('twitter_ads') %}
-{% else %}
-    {% do exclude_list.append('twitter_ads') %}
-{% endif %}
-
-{# Log which packages were added to which list #} 
-{{ log("\n Include list before filtering: " ~ include_list, info=True) }}
-{{ log("\n Exclude list before filtering: " ~ exclude_list, info=True) }}
-
-{% set enabled_packages = get_enabled_packages(exclude=exclude_list) %}
-
-{# Log final enabled packages after filtering by global enablement #}
-{{ log("\n Enabled packages after get_enabled_packages(): " ~ enabled_packages, info=True) }}
 {{ config(enabled=is_enabled(enabled_packages)) }}
 
 with